DDCNLS-7107 | Added integration test for caching
